chore: extracted FeatureFlags for cloud usage
