+++ pkg/agent/agent_test.go
@@ -20,9 +20,9 @@ import (
 	"google.golang.org/grpc/metadata"
 
 	"github.com/kubeshop/testkube/internal/config"
-	"github.com/kubeshop/testkube/internal/featureflags"
 	"github.com/kubeshop/testkube/pkg/agent"
 	"github.com/kubeshop/testkube/pkg/cloud"
+	"github.com/kubeshop/testkube/pkg/featureflags"
 )
 
 func TestCommandExecution(t *testing.T) {
