+++ pkg/agent/events_test.go
@@ -18,10 +18,10 @@ import (
 	"google.golang.org/grpc/metadata"
 
 	"github.com/kubeshop/testkube/internal/config"
-	"github.com/kubeshop/testkube/internal/featureflags"
 	"github.com/kubeshop/testkube/pkg/agent"
 	"github.com/kubeshop/testkube/pkg/api/v1/testkube"
 	"github.com/kubeshop/testkube/pkg/cloud"
+	"github.com/kubeshop/testkube/pkg/featureflags"
 )
 
 func TestEventLoop(t *testing.T) {
